src/app/views/add/add.component.spec.ts
import {ComponentFixture, TestBed} from '@angular/core/testing';

import {AddComponent} from './add.component';
import {provideHttpClient} from '@angular/common/http';
import {provideAnimations} from '@angular/platform-browser/animations';
import {provideRouter} from '@angular/router';
import {By} from '@angular/platform-browser';

describe('AddComponent', () => {
  let component: AddComponent;
  let fixture: ComponentFixture<AddComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [AddComponent],
      providers: [provideRouter([]), provideHttpClient(), provideAnimations()]
    })
      .compileComponents();

    fixture = TestBed.createComponent(AddComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should have form', () => {
    const form = fixture.debugElement.query(By.css('form'));
    expect(form).toBeTruthy();
  });

  it('should have empty category control', () => {
    const form = fixture.debugElement.query(By.css('form'));
    const input = form.query(debugEl => debugEl.nativeNode.id === 'category');
    expect(input.nativeElement.value === '').toBeTruthy();
  });
});
